weight_col as optional param, use floats not ints in ipf
This checkin offers solutions to two issues:

1. The weight column was hard-coded as 'total'. This is now the
default value for an optional parameter, weight_col, to the ipfn
methods.

2. If the weights in the weight column are of type int64 and fairly
large, it is possible to overflow the data type during raking,
which fails quietly and writes garbage data to the Pandas series.

To avoid this, convert data type to float with .astype() Series method