Avoid unnecessary copies
The self.original and self.aggregates are never modified but kept the
same as the original user input.  The `m` and `aggregates` variables are
a copy of the user input and may be modified during the iterations.

The `iteration` method is responsible for converting the user input to a
numpy array of floats.  The `ipfn_np` function only checks whether that
condition holds but does no conversion.  This allows the `ipfn_np` to be
used directly for user that want absolute minimum memory usage without
any extra copy of the input array.

Instead of always converting to float64, the input arrays are now only
checked if they are of floating point type.  This means that a single
precision float32 array will be kept that way, allowing for a smaller
memory footprint, if the problem allows the lower precision.